Remove set_latency_target(), add lock_memory_to_ram()
set_latency_target() attempts to set keep "low latency" by
- writing a specific frequency to /dev/cpu_dma_latency
- setting manually the minimum CPU processor frequency to be the maximum

There is no functionality to undo this, and most users are probably not
even aware that OAI does this. It is generally preferable to set this
beforehand using a governor or by disabling sleep states (as outlined in
the tutorial, and in a follow-up commit in the performance tuning docs).

The previous mlockall() call, to lock memory to RAM, is retained in a
new function.